+++ commandIns/init-Configs/gitignore.js
@@ -1,0 +1,14 @@
+module.exports = {
+    path: '.gitignore',
+    content: `
+/node_modules
+
+/**/node_modules
+
+/**/build
+/**/dist/
+# misc
+.DS_Store
+
+npm-debug.log*`
+}