+++ config/variable.js
@@ -1,13 +1,24 @@
-// 自定义命令目录名称
-const CUSTOM_DIR = 'weiyi-tools'
-const CUSTOM_SHELL_DIR = 'weiyi-tools-shell'
+// 应用内自定义命令目录名称
+const A_CUSTOM_COMMAND_DIR = 'custom-commands'
+// 用户端自定义命令目录名称
+const C_CUSTOM_COMMAND_DIR = `weiyi-tools-${A_CUSTOM_COMMAND_DIR}`
+// 应用内自定义snippet目录名称
+const A_CUSTOM_SNIPPETS_DIR = 'custom-snippets'
+// 用户端自定义snippet的目录名称
+const C_CUSTOM_SNIPPETS_DIR = `weiyi-tools-${A_CUSTOM_SNIPPETS_DIR}`
+// 应用内initConfig 的仓库目录名称
+const A_INITCONFIG_DIR = 'init-Configs'
+
 // 自定义命令接入文档地址
 const ACCESS_DOCUMENT_URL = 'https://wedog.yuque.com/vg2ro2/rcdfga/rionk7'
 
 
 
 module.exports = {
-    CUSTOM_DIR,
-    ACCESS_DOCUMENT_URL,
-    CUSTOM_SHELL_DIR
+    C_CUSTOM_COMMAND_DIR,
+    A_CUSTOM_COMMAND_DIR,
+    C_CUSTOM_SNIPPETS_DIR,
+    A_CUSTOM_SNIPPETS_DIR,
+    A_INITCONFIG_DIR,
+    ACCESS_DOCUMENT_URL
 }